+++ lib/monolens/error.rb
@@ -1,4 +1,11 @@
 module Monolens
-  class Error < StandardError; end
-  class LensError < Error; end
+  class Error < StandardError
+  end
+  class LensError < Error
+    def initialize(message, location = [])
+      super(message)
+      @location = location
+    end
+    attr_reader :location
+  end
 end