+++ lib/monolens/object/transform.rb
@@ -12,25 +12,27 @@ module Monolens
       end
 
       def call(arg, world = {})
-        is_hash!(arg)
+        is_hash!(arg, world)
 
         result = arg.dup
         option(:defn, {}).each_pair do |attr, sub_lens|
-          actual_attr, fetched = fetch_on(attr, arg)
-          if actual_attr
-            result[actual_attr] = sub_lens.call(fetched, world)
-          else
-            on_missing(result, attr)
+          deeper(world, attr) do |w|
+            actual_attr, fetched = fetch_on(attr, arg)
+            if actual_attr
+              result[actual_attr] = sub_lens.call(fetched, w)
+            else
+              on_missing(result, attr, world)
+            end
           end
         end
         result
       end
 
-      def on_missing(result, attr)
+      def on_missing(result, attr, world)
         strategy = option(:on_missing, :fail)
         case strategy.to_sym
         when :fail
-          raise LensError, "Expected `#{attr}` to be defined"
+          fail!("Expected `#{attr}` to be defined", world)
         when :null
           result[attr] = nil
         when :skip