+++ lib/monolens/str/split.rb
@@ -4,7 +4,7 @@ module Monolens
       include Lens
 
       def call(arg, world = {})
-        is_string!(arg)
+        is_string!(arg, world)
 
         sep = option(:separator)
         sep ? arg.split(sep) : arg.split