+++ spec/monolens/coerce/test_datetime.rb
@@ -9,9 +9,34 @@ describe Monolens, 'coerce.datetime' do
     expect(subject.call('11/12/2022 17:38')).to eql(DateTime.parse('2022-12-11 17:38'))
   end
 
-  it 'fails on invalid dates' do
-    expect {
-      subject.call('invalid')
-    }.to raise_error(Monolens::LensError)
+  describe 'error handling' do
+    let(:lens) do
+      Monolens.lens({
+        'array.map' => {
+          :lenses => 'coerce.datetime'
+        }
+      })
+    end
+
+    subject do
+      begin
+        lens.call(input)
+        nil
+      rescue Monolens::LensError => ex
+        ex
+      end
+    end
+
+    let(:input) do
+      ['invalid']
+    end
+
+    it 'fails on invalid dates' do
+      expect(subject).to be_a(Monolens::LensError)
+    end
+
+    it 'properly sets the location' do
+      expect(subject.location).to eql([0])
+    end
   end
 end