+++ spec/monolens/object/test_select.rb
@@ -150,4 +150,29 @@ describe Monolens, 'object.select' do
       expect(subject.call(input)).to eql(expected)
     end
   end
+
+  describe 'error traceability' do
+    let(:lens) do
+      Monolens.lens('object.select' => {
+        defn: {
+          status: :priority
+        }
+      })
+    end
+
+    subject do
+      lens.call(input)
+      nil
+    rescue Monolens::LensError => ex
+      ex
+    end
+
+    let(:input) do
+      {}
+    end
+
+    it 'correctly updates the location' do
+      expect(subject.location).to eql([:status])
+    end
+  end
 end